Permissions for project author only

+++ app/controllers/project_users_controller.rb
@@ -1,24 +1,34 @@
 class ProjectUsersController < ApplicationController
   def destroy
-    p = ProjectUser.where(project_id: params[:project_id], user_id: params[:user_id])
-    if p.destroy_all
-      flash[:notice] = "Successfully deleted user from project."
-      redirect_to project_path(id: params[:project_id])
+    if current_user == Project.find(params[:project_id]).author
+      p = ProjectUser.where(project_id: params[:project_id], user_id: params[:user_id])
+      if p.destroy_all
+        flash[:notice] = "Successfully deleted user from project."
+        redirect_to project_path(id: params[:project_id])
+      else
+        flash[:error] = "Error occurred."
+        redirect_to project_path(id: params[:project_id])
+      end
     else
-      flash[:error] = "Error occurred."
+      flash[:error] = "Only project creator can delete users from project."
       redirect_to project_path(id: params[:project_id])
     end
   end
 
   def create
-    p = ProjectUser.new
-    p.project_id = params[:project_id]
-    p.user_id = params[:user_id]
-    if p.save
-      flash[:notice] = "Successfully added user to project."
-      redirect_to project_path(id: params[:project_id])
+    if current_user == Project.find(params[:project_id]).author
+      p = ProjectUser.new
+      p.project_id = params[:project_id]
+      p.user_id = params[:user_id]
+      if p.save
+        flash[:notice] = "Successfully added user to project."
+        redirect_to project_path(id: params[:project_id])
+      else
+        flash[:error] = "Error occurred."
+        redirect_to project_path(id: params[:project_id])
+      end
     else
-      flash[:error] = "Error occurred."
+      flash[:error] = "Only project creator can add users to project."
       redirect_to project_path(id: params[:project_id])
     end
   end
